feat: add clear with cascade option for cockroachdb

+++ src/driver/aurora-mysql/AuroraMysqlQueryRunner.ts
@@ -1966,8 +1966,12 @@ export class AuroraMysqlQueryRunner
      */
     async clearTable(
         tableOrName: Table | string,
-        options?: { cascade?: boolean },
+        options?: { cascade: boolean },
     ): Promise<void> {
+        if (options?.cascade)
+            throw new TypeORMError(
+                `MySql does not support clearing table with cascade option`,
+            )
         await this.query(`TRUNCATE TABLE ${this.escapePath(tableOrName)}`)
     }
