+++ src/driver/mongodb/MongoQueryRunner.ts
@@ -1382,8 +1382,13 @@ export class MongoQueryRunner implements QueryRunner {
      */
     async clearTable(
         collectionName: string,
-        options?: { cascade?: boolean },
+        options?: { cascade: boolean },
     ): Promise<void> {
+        if (options?.cascade) {
+            throw new TypeORMError(
+                `MongoDB driver does not support clearing table with cascade option`,
+            )
+        }
         await this.databaseConnection
             .db(this.connection.driver.database!)
             .dropCollection(collectionName)
