+++ src/driver/spanner/SpannerQueryRunner.ts
@@ -1534,8 +1534,13 @@ export class SpannerQueryRunner extends BaseQueryRunner implements QueryRunner {
      */
     async clearTable(
         tableName: string,
-        options?: { cascade?: boolean },
+        options?: { cascade: boolean },
     ): Promise<void> {
+        if (options?.cascade) {
+            throw new TypeORMError(
+                `Spanner does not support clearing table with cascade option`,
+            )
+        }
         await this.query(`DELETE FROM ${this.escapePath(tableName)} WHERE true`)
     }
