+++ src/driver/sqlserver/SqlServerQueryRunner.ts
@@ -2724,8 +2724,13 @@ export class SqlServerQueryRunner
      */
     async clearTable(
         tablePath: string,
-        options?: { cascade?: boolean },
+        options?: { cascade: boolean },
     ): Promise<void> {
+        if (options?.cascade) {
+            throw new TypeORMError(
+                `SqlServer does not support clearing table with cascade option`,
+            )
+        }
         await this.query(`TRUNCATE TABLE ${this.escapePath(tablePath)}`)
     }
